Fix missing attachments in email templates
Whenever GetEmailTemplateConfigsAsync() was called, it would only get the template config and not the attachments in the config. This change basically gets the metadata for all of the templates and calls GetEmailTemplateConfigAsync() to get everything for each individual template.

Ya know... I could have probably spotted that with unit tests. Should probably get on that... 🙃